style: black-format 3 branch-modified files for #3548 CI
Pure black 26.5.1 reformat (no semantic change) of the three files this
branch modifies vs main: aiter/fused_moe.py,
op_tests/multigpu_tests/test_dispatch_combine.py,
aiter/dist/device_communicators/communicator_pynccl.py.

Other repo-wide black drift from the main bump is intentionally left
untouched (not part of this PR's scope).

+++ aiter/fused_moe.py
@@ -1981,7 +1981,9 @@ def apply_act_and_mul(out, act_input, activation=ActivationType.No):
     else:
         inter_dim = act_input.shape[-1] // 2
         torch_act = aiter.get_torch_act(activation)
-        result = torch_moe_act(act_input, torch_act, inter_dim, activation).to(out.dtype)
+        result = torch_moe_act(act_input, torch_act, inter_dim, activation).to(
+            out.dtype
+        )
         out.view(-1, inter_dim).copy_(result.view(-1, inter_dim))
     return out
